refactor: extract PR display logic to pr_display.lua
Move build_pr_content, show_pr_info, and supporting buffer/window
functions to a dedicated pr_display module. lualine.lua now only
contains component, get_info, and on_click, with backward-compatible
re-exports of build_pr_content and show_pr_info.

Update tests to require from ghsigns.pr_display directly.
